requirements.txt
@"
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
alembic==1.13.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
waitress==2.1.2
gunicorn==21.2.0
sendgrid==6.11.0
requests==2.31.0
pytz==2024.1
python-dotenv==1.0.0
"@ | Out-File -FilePath "requirements.txt" -Encoding UTF8